+++ src/index.js
@@ -3,4 +3,4 @@ export { loadProjectConfig } from "./config/index.js";
 export { runAudit, runGlossaryCheck, runRuntime } from "./api/index.js";
 export { createReport, emitReport, formatResults, hasFailures } from "./reporting/index.js";
 export { buildRepositoryContext, coreRules, loadRulePacks, runRules } from "./rule-engine/index.js";
-export { bootstrapRuntime, inferProjectState, classifyChange, routeTask, resolveGates, applyDraftDocUpdates, syncNavigation, recordHumanDecision, readHumanDecisions, findLatestDecisionForGate, loadResumeInputFromReport, createRuntimeReport, formatRuntimeReport } from "./runtime/index.js";
+export { bootstrapRuntime, inferProjectState, classifyChange, routeTask, resolveGates, applyDraftDocUpdates, syncNavigation, recordHumanDecision, readHumanDecisions, findLatestDecisionForGate, getLatestGateDecisionMap, GATE_DECISION_STATUSES, normalizeGateDecisionStatus, buildGateDecisionMap, listApprovedGates, loadResumeInputFromReport, validateRuntimeReportShape, readRuntimeReport, writeRuntimeReport, RUNTIME_FINAL_STATUSES, isHumanDecisionRuntimeStatus, deriveRuntimeFinalStatus, createRuntimeReport, formatRuntimeReport } from "./runtime/index.js";